task2/backend/llm.py
import google.generativeai as genai
import os
from dotenv import load_dotenv
from .prompts import USER_RESPONSE_PROMPT, ADMIN_SUMMARY_PROMPT, RECOMMENDED_ACTION_PROMPT

load_dotenv()

class LLMService:
    def __init__(self):
        api_key = os.getenv("GEMINI_API_KEY")
        if not api_key:
            print("Warning: GEMINI_API_KEY not set.")
            self.model = None
        else:
            genai.configure(api_key=api_key)
            self.model = genai.GenerativeModel('gemini-pro')

    async def generate_response(self, prompt):
        if not self.model:
            return "LLM Service Unavailable"
        try:
            response = self.model.generate_content(prompt)
            return response.text.strip()
        except Exception as e:
            print(f"LLM Error: {e}")
            return "Error generating response"

    async def process_review(self, review_text: str, rating: int):
        """
        Generates all AI components for a review.
        """
        # Prepare prompts
        response_prompt = USER_RESPONSE_PROMPT.format(review_text=review_text, rating=rating)
        summary_prompt = ADMIN_SUMMARY_PROMPT.format(review_text=review_text)
        action_prompt = RECOMMENDED_ACTION_PROMPT.format(review_text=review_text, rating=rating)

        # In a production environment, you might want to run these in parallel (asyncio.gather)
        # For simplicity/robustness here, sequential awaits or just sequential calls if the library blocks.
        # google.generativeai sync client is blocking. We can use the async one or run in executor.
        # For this assessment, we'll assume standard sync calls in threads via FastAPI's default behavior, 
        # or just wrap them if we were being super strict about async.
        # To keep it simple and runnable:
        
        user_response = await self.generate_response(response_prompt)
        summary = await self.generate_response(summary_prompt)
        action = await self.generate_response(action_prompt)

        return {
            "ai_response": user_response,
            "ai_summary": summary,
            "ai_action": action
        }

llm_service = LLMService()
